requirements.txt
streamlit==1.21.0
torch==2.0.0
torchvision==0.15.1
transformers==4.26.0
numpy==1.24.3
Pillow==9.3.0
requests==2.28.2